Introduce `FilePath` component
